Separating search field elements, scripts, styles, and tests.
