+++ assets/styles/advanced/_layout.scss
@@ -44,14 +44,3 @@ $class: '.advanced-search';
     margin-top: 1rem;
   }
 }
-
-
-
-// *                * //
-// * 2.1. ADD FIELD * //
-// *                * //
-
-#{$class}__add-field {
-  margin-left: auto;
-  margin-right: auto;
-}
