+++ views/advanced/partials/search-fields/_add-field.erb
@@ -1,0 +1,1 @@
+<button class="hide__no-javascript button__ghost advanced-search__add-field">Add another field</button>
