+++ views/advanced/partials/search-fields/_booleans.erb
@@ -1,13 +1,15 @@
-<fieldset class="advanced-search__search-field--booleans">
-  <legend class="visually-hidden">Select a boolean operator for field <%= index %></legend>
-  <%
-    booleans = @presenter.search_options.booleans
-    booleans.each do |boolean|
-      input_id = "#{boolean}-#{index}"
-  %>
-    <div class="checkbox">
-      <input type="radio" id="<%= input_id %>" name="boolean-<%= index %>" value="<%= boolean %>" <%= 'checked' if boolean == booleans.first %>>
-      <label for="<%= input_id %>"><%= boolean %></label>
-    </div>
-  <% end %>
-</fieldset>
+<% if index > 0 %>
+  <fieldset class="advanced-search__search-field--booleans">
+    <legend class="visually-hidden">Select a boolean operator for field <%= index %></legend>
+    <%
+      booleans = @presenter.search_options.booleans
+      booleans.each do |boolean|
+        input_id = "#{boolean}-#{index}"
+    %>
+      <div class="checkbox">
+        <input type="radio" id="<%= input_id %>" name="boolean-<%= index %>" value="<%= boolean %>" <%= 'checked' if boolean == booleans.first %>>
+        <label for="<%= input_id %>"><%= boolean %></label>
+      </div>
+    <% end %>
+  </fieldset>
+<% end %>
