+++ views/advanced/partials/search-fields/_remove-field.erb
@@ -1,0 +1,6 @@
+<% if index > 0 %>
+  <button type="button" class="hide__no-javascript button__ghost advanced-search__remove-field" data-field-id="<%= search_field_id %>">
+    <span class="material-symbols-rounded" aria-hidden="true">close</span>
+    <span class="visually-hidden">Remove search field <%= index %></span>
+  </button>
+<% end %>
